requirements.txt
domdf-python-tools>=1.7.0
dulwich>=0.20.14
flask>=1.1.2
flask-githubapp>=0.2.0
flask-sqlalchemy>=2.4.4
github3-py>=1.3.0
gunicorn>=20.0.4
repo-helper>=2020.11.22
southwark>=0.6.0
